requirements.txt
pyautogui
pytesseract
Pillow
openpyxl
opencv-python
pyperclip
